requirements.txt
playwright
PyQt6
pandas
openpyxl
pyyaml
qasync